add boost license for hash_combine, move to diff header
