+++ cpp/include/cudf/types.hpp
@@ -27,7 +27,6 @@
 #include <cassert>
 #include <cstddef>
 #include <cstdint>
-#include <functional>
 #include <iterator>
 
 /**
@@ -329,22 +328,6 @@ inline bool operator!=(data_type const& lhs, data_type const& rhs) { return !(lh
 std::size_t size_of(data_type t);
 
 /**
- * @brief Combines two hashed values into a single hashed value.
- *
- * Adapted from Boost hash_combine function, modified for 64-bit
- * https://www.boost.org/doc/libs/1_35_0/doc/html/boost/hash_combine_id241013.html
- * https://stackoverflow.com/a/4948967/1550940
- *
- * @param lhs The first hashed value
- * @param rhs The second hashed value
- * @return Combined hash value
- */
-constexpr std::size_t hash_combine(std::size_t lhs, std::size_t rhs)
-{
-  lhs ^= rhs + 0x9e3779b97f4a7c15 + (lhs << 6) + (lhs >> 2);
-  return lhs;
-}
-/**
  *  @brief Identifies the hash function to be used
  */
 enum class hash_id {
@@ -362,15 +345,3 @@ static constexpr uint32_t DEFAULT_HASH_SEED = 0;
 
 /** @} */
 }  // namespace cudf
-
-// specialization of std::hash for cudf::data_type
-namespace std {
-template <>
-struct hash<cudf::data_type> {
-  std::size_t operator()(cudf::data_type const& type) const noexcept
-  {
-    return cudf::hash_combine(std::hash<int32_t>{}(static_cast<int32_t>(type.id())),
-                              std::hash<int32_t>{}(type.scale()));
-  }
-};
-}  // namespace std
